requirements.txt
beautifulsoup4==4.13.4
certifi==2025.6.15
charset-normalizer==3.4.2
chess==1.11.2
colorama==0.4.6
idna==3.10
matplotlib==3.8.4
numpy==2.3.1
pandas==2.3.0
python-chess==1.999
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.4
seaborn==0.13.2
six==1.17.0
soupsieve==2.7
stockfish==3.28.0
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
